clients/python/kernel-memory-client/pyproject.toml
[tool.poetry]
name = "kernel-memory-client"
version = "1.0"
description = "A client library for accessing Microsoft.KernelMemory.ServiceAssembly"
authors = []
readme = "README.md"
packages = [
    {include = "kernel_memory_client"},
]
include = ["CHANGELOG.md", "kernel_memory_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
